requirements.txt
pandas>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.15.0
googlemaps==4.10.0
google-search-results>=2.4.0
lxml>=4.9.0
python-dotenv==1.0.1
webdriver-manager>=4.0.0
fake-useragent>=1.4.0
geopy>=2.3.0
python-slugify>=8.0.0
neo4j>=5.14.0
tqdm>=4.66.1
pymongo==4.6.1